Turn debugging on and off via xml file
Check to see whether debugging is turned on or off before we start writing it out to the console.  If it's off then don't do anything, if it is on then debug as required.  This means we don't have to amend the code manually and we can set the value of debug in the config file outside of the program code once released.